Refactor rate limiting implementation into dedicated extension methods
